feat(aa): add ESM support
This adds ESM support to `aa` insofar as it will inspect the AST for imports/exports.  It does not support `await import()`.

It also accepts a custom resolver, which is needed for actually resolving ESM packages, due to the inability of the `resolve` package to do so.  Adding this natively would break the custom resolver; we may need to abandon it, because `package.json` _must_ be read in order to resolve any package having an `exports` field.